+++ src/parse/setup/setup_redir.c
@@ -6,29 +6,12 @@
 /*   By: mschlenz <mschlenz@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/14 21:31:35 by mschlenz          #+#    #+#             */
-/*   Updated: 2022/10/30 16:45:47 by mschlenz         ###   ########.fr       */
+/*   Updated: 2022/10/30 17:34:46 by mschlenz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
-// static bool	setup_redir_files(t_data *data, char *cmd, int *i)
-// {
-// 	if (!data->flags->redir_out)
-// 	{
-// 		printf("hello\n");
-// 		if (!setup_filenames(data, i, cmd, 0))
-// 			return (true);
-// 	}
-// 	else
-// 	{
-// 		if (!setup_filenames(data, i, cmd, 1))
-// 			return (true);
-// 	}
-// 	return (false);
-// }
-
-// data->flags->redir_in = false;
 static bool	setup_redir_out(t_data *data)
 {
 	int		tmp_fd;
@@ -47,7 +30,7 @@ static bool	setup_redir_util(t_data *data, char *cmd, int *i)
 {
 	if (setup_argv_set_redir_flags(data, cmd, i))
 		return (true);
-	if ((*i) < ft_strlen(cmd))
+	if ((*i) < (int)ft_strlen(cmd))
 	{
 		(*i)++;
 		if (setup_pipe(data, cmd, i))
